requirements.txt
# --- Core dependencies ---
mysql-connector-python>=9.5.0
python-dotenv>=1.2.1
requests>=2.31.0

# --- Optional for data processing ---
pandas>=2.2.2

# --- For logging and JSON handling ---
rich>=13.9.1
# requirements.txt additions (append these lines)
